Rename parse.py to util.py
